scripts/requirements.txt
mysql-connector-python==8.1.0
bcrypt==3.2.0